pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invoice-extraction"
version = "0.1.0"
description = "Extract structured invoice data from PDFs/images using Gemini AI"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.35.0",
    "langchain-google-genai>=4.0.0",
    "langchain-core>=0.3.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
